astro: refactor
