+++ packages/example-floors-astro/src/content.config.ts
@@ -2,7 +2,7 @@ import { file } from 'astro/loaders'
 import { defineCollection } from 'astro:content'
 
 import { collections as collectionsCommon } from '../node_modules/svgmapviewer-astro-floors/src/content.config'
-import { xinfoSchema } from './utils/schema'
+import { xinfoSchema } from './schema'
 
 const pois = defineCollection({
   loader: file('./src/content/pois.yaml'),
