+++ packages/example-floors-astro/src/schema.ts
